[refactor] 메시지 변환 담당 클래스 변경
