+++ src/main/java/com/jaeyeonling/racingcar/domain/Participants.java
@@ -1,9 +1,12 @@
 package com.jaeyeonling.racingcar.domain;
 
+import com.jaeyeonling.racingcar.utils.StringUtils;
+import com.jaeyeonling.racingcar.view.Visualizable;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
-class Participants {
+class Participants implements Visualizable {
 
     private final List<Car> participants;
 
@@ -17,4 +20,20 @@ class Participants {
     List<Car> getCars() {
         return participants;
     }
+
+    @Override
+    public String visualize() {
+        return getVisualGameStatus();
+    }
+
+    private String getVisualGameStatus() {
+        final StringBuilder visualBuilder = new StringBuilder();
+        for (final Car car : getCars()) {
+            visualBuilder.append(car.visualize())
+                    .append(StringUtils.NEW_LINE);
+        }
+
+        return visualBuilder.toString();
+    }
+
 }
